♻️ Separate test and test cases, helper function  (#1346)
