+++ src/test/lib/common/test_helpers.ts
@@ -1,0 +1,5 @@
+import { test } from 'vitest';
+
+export function runTests<T>(testName: string, testCases: T[], testFunction: (testCase: T) => void) {
+  test.each(testCases)(`${testName} - %o`, testFunction);
+}
